refactor: 刪除 RecipeCard 元件並更新引用路徑至 common 目錄以改善檔案組織。
